drouting: Fix an SHM leak with "acc_call_params_avp"
Do NOT store an SHM pointer into an AVP, as you have a minimal chance of
somehow getting invoked so you can properly free it!  The solution is to
store the entire struct into the AVP, as a (char *) buffer -- the tm
module will properly free the whole AVP (including this buffer)
afterwards.

Thanks to Răzvan for the tip!