Added AUTH_JWT module
In some cases ( ie. WebRTC ) the user authenticates on another
layer ( eg. on the actual website ), so it makes no sense to
double-authenticate it on the SIP layer.
Thus, the WebRTC client will simply present the JWT auth token
it received from the server, and pass it on to
OpenSIPS ( ie. Authorization: Bearer jwt_token_here )
which will use that for authentication purposes.