dialog: do not update topo hiding advertised contact
This commit fixes a bug introduced by commit 7b7a9fc2bf, that was
updating the contact *after* topology hiding has replaced it. This bug
would be triggered when a re-INVITE/UPDATE was triggered on a call that
had topology hiding enabled - in these scenarios OpenSIPS would re-learn
the contact advertised by topology hiding, that is OpenSIPS itself,
resulting in looped sequential requests.