tls_mgm: enforce at least TLSv1 when SSLv2.3 is used
Starting with Debian 10 (buster), the OpenSSL library is compiled with
TLSv1.2 as the minimum TLS version. Therefore, when specifying SSLv2.3,
basically only TLSv1.2 and TLSv1.3 can be used.
This commit sets the minimum version to TLSv1, since SSLv2.3 will soon
be deprecated.

Thanks go to @telematico for reporting this issue
Close #1986