The main process (attendant) is no longer running the child_init()
The main proc must stay free of any db connections, so it can later fork and child_init() other procs (due auto_scaling).
For the modules which were keeping a DB conn open in main proc for later usage, during module destroy, the new approach is to open the conn, on the spot, in the destroy handler.

Related to #2003