HP_MALLOC: Add "re-scanning" logic
Given that HP_MALLOC has fine-grained locking, if the big fragment
shifts down from hash bucket N to bucket N-1 due to another
process performing the allocation, we may actually "lose" it during
our own scan, since bucket N-1 was empty and we're now waiting
for bucket N to unlock.  When it does unlock, it will also be empty.

As a solution: retry the scan up to N times, as long as it's feasible!

Many thanks to 46Labs for supporting this work!