dialog: fix a replicated profiles counting issue
The incorrect profile counting was caused by a race between the processing of
the replicated dialog create/update packets. While processing the create
packet, the lock was released before linking the dialog into a profile and
setting the sharing tag. As such, the process that handled the update packet
could insert the dialog into a profile itself, but with an improper sharing tag.

Thanks to 46Labs for reporting and help with troubleshooting this!

(cherry picked from commit 7e623b302191b6a34087b1527ef5383914ed1663)