HP_MALLOC: Aggregate SHM statistic updates
It is better to update the stats with a diff of the
attached/detached/split/merged fragment changes.  Updating the stats in
a "step by step" manner has the side-effect of creating temporarily
bogus values. (i.e. negative "used" and/or "real_used")

Many thanks to 46Labs for supporting this work!