This change causes epoll_wait to only return events for a single FD on everycall.
With the use of EPOLLEXCLUSIVE we want to ensure that a single process is not being
woke to handle multiple FDs at once. This allows EPOLL to better distribute wake-ups to
processes that are actually ready to run. Without this a single process may be woken to handle
IO for multiple FDs, while other processes are available waiting for work.