Fix un-released transaction upon t_wait_for_new_branches + cancel.
If t_wait_for_new_branches() is used and no real signaling branch is added, the caller canceling must not be ignored upon branch selection. The injected 408 timeout on the phony branch must be considered, leading the transaction termination.

Thanks to Chris Maciejewski for reporting and providing troubleshooting info.
Also thanks to Razvan Crainea ( @razvancrainea ) for investigating and reproducing the problem.