[call_center] pass via cc_handle_call() a custom param to B2B scripts
an opaque string to be passed as parameter to the "callcenter" and "agent" B2B scenarios. It is itended for custom integration of the call center module and it is 100% up to the script writer about the value and purpose of this parameter, OpenSIPS will not touch or iterpret it.