b2b_logic: replicate entire tuple info for all entity create events
Even though it would be optimal to include the entire tuple info in a single
replicated entity, this would cause races on the receiving instance when
processing the packets. New entities carrying only tuple "updates" could be
processed before actually creating the tuple. As such, the b2b_logic level
entity could not be created in these cases.