cachedb_mongodb: fix crashes when loading module along with tls_mgm
Loading the cachedb_mongodb module causes the openssl library to be
initialized in the pre-daemon process, by the mongoc library's
constuctor, before tls_mgm gets the chance to do its own initialization
in the attendant process.
This commit overwrites the openssl functions used by the mongoc library
in order to actually skip the initialization and leave it to tls_mgm.

Related to #2091