siprec: send BYE for sessions not established
When a SIPREC session was started on a 183 reply, but the session was
not stopped, a BYE to the SIPREC server should be sent to close the
session.

Thanks go to Krunal Patel for reporting this and offering the fix.

Close #2136

(cherry picked from commit b4cd132731f8e2fc0a949ea52b22952879b345ca)