Adjusted TCP conenction search algorithm
A previous commit that considers "send_sock" when searching for a TCP connection was implemented in a way that fell-back to just a destination ip:port search when the source+destination ip:port search failed to return a connection.  It became evident later that this algorithm was also flawed because it could return an incorrect connection (one for a different source ip:port) when it failed to match a connection with the specified source ip:port because the correct connection had recently been disconnected. The algorith was adjusted to *only* search using both source+destination ip:port whenever a valid "send_sock" is specified, and only use the prior destination-only ip:port when "send_sock" is not specified.