fraud_detection: Add the possibility to disable stats
A (0) value for any of the fraud detection thresholds now means: "ignore
this threshold", instead of: "raise an event on each call for this
threshold?!".  Although (0) values could, in theory, have been useful
for event generation testing purposes before this patch, this is a small
price to pay in order to gain the ability to disable a stat.  Developers
can now add 1 more CPS and still keep testing with a new (1) value :)

Although the DB schema changes slightly, due to the added "DEFAULT 0",
it is fully backwards-compatible with the previous one, except
fraud_detection row INSERTs will now be more quiet, generating no more
warnings from the backend (e.g. "Column does not have a default
value..." -> but still defaults to 0, thanks to non-strict mode, leading
to strange warnings from the module).

Fixes #890