cfgutils: Enhance check_time_rec() with multi-rec support
This gives developers the ability to combine multiple time recurrences
into a compact, easy-to-evaluate string expressing a complex schedule.
The linking is done using the logical operators: AND ("&"), OR ("/")
and NEG ("!").

Assuming A, B and C are vanilla check_time_rec() time recurrence
strings, here are some example expressions that now become available:
    A & B
    A / B
    (A & B) / C
    A & !B
    !((A / B) & C) / D