sipcapture: fix overlapping function name
the build_dummy_msg() function being used without the static identifier
causes confusion for some compilers. An example is the AL2 compiler,
which instead of using the module's implementation, it uses the core's
one, hence resulting in broken behavior.

Credit goes to @ghmj2417 (GitHub) for reporting this in ticket #2255
Close #2255