topo_hiding: always allocate space for DID in user
When contact could not be parsed, the new Contact's prefix was not
allocating bytes in buffer for storing for the dialog ID in the user
part. However, the DID was appended to the prefix without any error
checks, resulting in a memory corruption. This commit always allocates
spaces for storing the Dialog ID, even if the contact could not be
parsed.
This fixes all uses of the topology_hiding() function with the `D` flag.

Thanks a lot Sergei Lavrov (@ccppprogrammer) for reporting this in
ticket #2285 and for offering information for figure it out. Close #2285