tls: fix certificate matching when reusing connections
The SSL_CTX pointers may not be equal for the same SSL pointer (connection)
after the rework that changed the storage of the context to be
per-process. The tls_domain saved in the SSL extra storage will now be
used for matching the certificates instead.

(cherry picked from commit 582170270c3e0e41d539108b640552ae3cb23cfd)