tls_mgm: add API functions for operations with a SSL_CTX
Add API functions to set the CAs, certificate and private key for a given
SSL_CTX (with a different SSL_CTX, or a file in case of private keys, as a
source). These functions can be used from modules which require some TLS
operations but without having to link with openssl directly.