ratelimit: keep only local values in pipe->counter for SBT
Up until this commit, the SBT algorithm would keep in pipe->counter the
sum between its local value, as well as the values gathered from the
other servers in the cluster. This means that it would replicate the
other's counters as well, leading to a broken behavior.
This fix makes sure that the SBT algorithm only stores the local counter
in the pipe->counter field, therefore it will not replicate other's
counters.

(cherry picked from commit 9621cac2033308e1d386103a229702750ffb005b)