dialog: load from DB call looping dialogs
When loading a dialog from DB that has the same coordinates (same
callid+to-tag+from-tag pair) with a different dialog in memory, we
should not ignore it if it comes with a different `hash_id`, as it might
be a completely different dialog, resulted from a call looping scenario
(call that passes through the same OpenSIPS instance twice), which
legitimately creates a new dialog with a new `hash_id`.

Thanks go to @perwx3 (GitHub) for reporting it in ticket #2311.

Complements commit b26d59e3.
Fix #2311.