dialog: store dialog in transaction in dlg_onroute
When the transaction is created before matching the dialog, there is was
no callback to store the dialog in the transaction - as a consequence,
the acc cdr callbacks were not able to find the dialog, thus no longer
generate dialogs. This commit fixes this bevavior.

(cherry picked from commit 0043bb0e6084e60fbda88269c0ddd8f0301444fd)