Add const_str() macro, a better way to generate point-of-use string literal of type const str_const * to be used in cases like:
    str_strcmp(in, _str("header"))

The problem with the _str() usage in this scenario among many
is that it updates the static pointer burried somewhere in
application guts with length and location of the buffer containing
its argument (e.g. "header" in this case) each time the
operation is performed and that extra work cannot be optimized
away. I have confirmed it with the Compiler Explorer, it costs
us two memory writes each time str_strcmp() is invoked.
Which not only isn't secure but also should have at least some
negative performance effect.

Use of:

     str_strcmp(in, const_str("header"))

in this scenario will not only be free of that effect, but also
pointer provided by the const_str can generally be passed
along and saved as needed, it will always be valid as long as
the application runs.