proto_tls/wss: fix crashes when dumping the openssl error stack
This commit serializes the execution of openssl's connect/accept/read/write
operations in order to prevent adding/removing entries concurrently to/from
the openssl error stack. If the performance penalty is deemed too high, the
NO_SSL_GLOBAL_LOCK compilation flag can be used to disable this behavior and
retain the risk of crashes.

Reported in #2362
Credits to Alexey Vasilyev for helping troubleshoot this.