rtpproxy: alocate anonymous structures in local scope
RTPP_CMD_IOVEC and RTPP_VCMD_INIT_STATIC are initializing a variable
with the content of another (anonymous one). If we however create the
anonymous variable in a block's scope, it will be unavailable at the
end of the block, resulting in invalid memory access - allocating the
structures in the function's context (rather than a block's context)
fixes this issue.
Bug introduced in 0d7a71b2.