Fix cleanup of the re-used dummy sip msg.
Be sure to clean ALL the fields in the msg, not only the ones related to the msg parsing. The prevent inheriting of data between sequential usage of the message, we need to clean the flags, sockets, etc.

(cherry picked from commit 4a2183a620c2ede9481b4369a459e2cf65dcaa48)