-d binary option: Evaluate before PKG malloc initializations
There is no need to delay the evaluation of '-d'.  By moving it earlier,
we gain the ability to control the logging of PKG memory initialization.

This patch doesn't change the fact that '-d' is somewhat broken,
because it will immediately get overridden by the 'log_level'
configuration file option.  But this is subject to a future patch.