stir_shaken(): Fix too relaxed E.164 conditions; Add number checks
Commit 8c842cc33d made it such that any string is accepted as valid
STIR/SHAKEN identity, in "loose mode".  For example,
"From: sip:foobar@test.com" would pass as valid CLI and "foobar" would
get written in the PASSporT, which cannot be OK.

This commit also makes it so the new "e164_strict_mode" modparam makes
the leading "+" mandatory in the From/To URIs.  Otherwise, the "+" is
optional and will get discarded if necessary.

Moreover, by moving the number checks around a bit, this commit also
fixes a bug where the opensips.cfg inputted orig/dest numbers are never
checked.  For example, if the developer were to have a bug where some
$var(cli) contains a SIP URI and gets passed to stir_shaken as
originator, the module would happily write the URI into the PASSporT as
"orig" and return success.