dialog: add event_id_format param
This parameter is used to specify the format of the dialog identifier in
the `E_DLG_STATE_CHANGED` event. It was added to prevent the event's
compatibility with previous 3.1 versions, but also allow the possibility
to have the id in a usable (hex or DB) form.
This is sort of backport of e3221aa1 and a fix of #2463