stir_shaken: Further improve commit 9c9cce6f36
It's not enough to check for missing PASSporT payload fields, as the
code can also crash on a cJSON type mismatch (malicious input?), e.g.:

	{... "orig": {"tn": ["1234"]} ...}

Here, the "orig-tn" is incorrectly given as a list, where it should be a
string.  Without an extra check for a NULL cJSON string subfield, the
code would still crash.

Many thanks to Sandro Gauci (Enable Security) for the find!

Issue discovered during OpenSIPIt'01 (https://opensipit.org/)