stats: deprecate unsupported_methods statistic
Removing the statistic as it is not reliable, due to the fact that
the term "unsupported_methods" is quite generic and does not reflect
the actual methods supported by OpenSIPS' script.
This commit simply prevents updating the statistic, as it is not
reliable.
Reported by Ben Newlin (@bcnewlin) in Ticket #2460

(cherry picked from commit 6a7e094eb17ec13d2ce94bb576b93b1c00f22511)