Migrate tcp conn ID from int to unsigned int
As used to generate array indexes (for tcp partitions), the ID of the TCP conn must be a positive number.
Still it was initialized with rand() (so possible high valules) and continously incremented -> this may lead to an overflow, resulting into a negative value (for an int data type).
The solution was to move to unsigned int, so we are safe when overflowing (still resutling into a positive number)
Fixes #2344

(cherry picked from commit 6a89cb8eb1f62578a71fa43353c564a7ea69c685)