rtpproxy: fix initializing child
If a process is not started with flag PROC_FLAG_INITCHILD, OpenSIPS will
not be interested in its init status, thus the extra process will not be
able to report its initializing state.
If however it was doing that, it would hit the bug reported in
ticket #2259, where the attendant was badly couting the number of
processes that need to be initialized, due to the fact that it would not
see any updates (in this case of the exports.procs structure) the
process would be doing in its mod_init().
Credits go to @ghmj2417 for reporting this in #2259 and to Bogdan
Andrei IANCU (@bogdan-iancu) for brainstorming the final solution.

(cherry picked from commit 6a920dd3fd1e5605d08acbd79b0b9e9994008d00)