sipcapture: establish the procs number at parsing
Updating the number of extra procs needed by the module in mod_init() is
a bad idea, since the attendant does not see any changes made by the
process that runs mod_init, thus it waits for more children tha it
should have, resulting in some bad pipe errors.
Credits go to @ghmj2417 for reporting this in #2259 and to Bogdan
Andrei IANCU (@bogdan-iancu) for brainstorming the final solution.