ratelimit: repurpose the repl_buffer_threshold
Instead of defaulting it to MTU (due to an old datagram limitation), we
should use a higher value, that would only control the fragmentation of
the send buffer.
Moreover, do not send the buffer to the clusters nodes under lock,
since it might block, thus block the entire ratelimit function.