db: Add IS NULL and IS NOT NULL unary operators
If you try to (mysql) prepare a statement with "... column IS ?" you'd
get an error.

For mysql, you could fix this by using the "... column <=> ?" special
operator ('IS NOT DISTINCT FROM'), but that's not standard SQL.

Instead, you can now do:

    update_cols[0] = &some_column;
    update_ops[0] = OP_IS_NULL;
    update_vals[0].nul = 1;

You still need to set .nul=1 because prepared statements have to consume
a single value.