dialog pinging: Avoid nested pinging transactions
This fixes a rare condition that could occur whenever the
OPTIONS/Re-INVITE pinging interval was lower than the SIP transaction
timeout (typically 15 seconds, so a pinging interval of < 15 seconds).

In this case, a new pinging transaction would be continuously started,
before the previous one would get the chance to complete, e.g. with 408
Timeout for a dead SIP endpoint.

Fixes #2535