dialog: Fix a subtle bug in parsing dialog IDs from DB
This fixes a rare issue where the dlg_parse_db_id() macro would
right-shift a signed "long long" value, potentially leading to
implementation or compiler defined behavior.

The fact that there is a binary diff after applying this patch proves
that there was an underlying, hard-to-detect issue in there:

-   4b192:	48 c1 f8 20          	sar    $0x20,%rax
+   4b192:	48 c1 e8 20          	shr    $0x20,%rax

Credits to Nick Altmann for providing some hints on this issue!
Related to #2504