Memory allocators: Improve the memory alignment support
This patch fixes the support for configurable memory alignment, such
that the ROUNDTO macro is now truly customizable, for any allocator.  As
long as the ROUNDTO is a multiple of 2, the allocators will return
memory aligned to any required multiple.

The defaults are still the same, as follows:
    * F_MALLOC: 8-byte aligned memory addresses
    * HP_MALLOC: 8-byte aligned memory addresses
    * Q_MALLOC: 16-byte aligned memory addresses (the default in libc!)

Also extend the allocator testing suite with memory alignment tests.
Related to #2546