b2b_entities/logic: fix a race condition with simultaneous reINVITEs
Fix behavior when receiving reINVITES from the connected entities at the
same time, one replied with 491 and the other with 200OK. The 200 OK
would not be passed to the other entity and the tuple would never be
deleted in this case.

Thanks to David Escartin from Sonoc for reporting and testing the fix.

(cherry picked from commit 9c9ab8f948be65dc6a99024f901e748d629f8e56)