Add code to troubleshoot some BUG report on weight sorting alg.
Dump all the gws, their weights and the running sums, so we can undestand why the alg fails here - this dump will happen only when the bug is detected.
Also, in case of this BUG, avoid dropping the call (by return -1), but rather return something usefull, so the call can be routed further.