ratelimit: proper expire for replicated pipes
Instead of updating the last_used on every received pipe, we shall only
consider the last locally used value when we want to expire. This way,
we only replicate pipes that have been locally used recently, if the
ones that are not, will not be replicated, thus will eventually be
deleted by the backup nodes.

Thanks go to Ken Rice (SIPNav) for reporting this and troubleshooting
the problem.