tls_wolfssl: fix behavior of is_peer_verified() with session tickets
If TLS session tickets were used for session resuming, the
is_peer_verified() script function would not be able to verify the peer
even if it did present a valid certificate in the initial TLS handshake.

Even so, this fix can only guarantee that the peer can be verified when
resuming a session, if the TLS domain is configured to require a peer
certificate initially. Otherwise, wolfssl does not provide a way of
retrieving the peer certificate from the received session ticket.

Fixes #2541