proto_bin: fix crash when parsed data is printed on error
The module was trying to print the parsed data in case of an error,
however it was using the `start` field of the tcp_conn. The problem is
that the module does not use the `start` field, therefore when a request
is not entirely read from the begining, it is not updated to point to
the new structure. Therefore, when we're trying to print the buffer, it
is pointing to an address on stack.
The fix for this bug was to suppress the parsed printing, as it is
binary anyway and it is the same value as the `buf`.

Many thanks to Bernard Buitenhuis for reporting the issue.