db_postgres: fix compilation for libpq versions < 9.3
Remove usage of the PQconninfo() function since it is not available for libpq
versions < 9.3. Conditionally enabling the code based on the libpq version is
not an option since there is no such macro available in libpq-fe.h.
PG_VERSION_NUM for example might not be present and refers to the
server version anyway.