Fix the DB saved description of the caller/callee sockets
Use the internal description of the socket when saving in DB. This means using the socket TAG when defined, versus the proto:ip:port description.
This is how the replication is also doing, to cope with potential changes of the socket IP during a replication. And in our case, during a restart.