Fix the DB saved description of the caller/callee sockets
Use the internal description of the socket when saving in DB. This means using the socket TAG when defined, versus the proto:ip:port description.
This is how the replication is also doing, to cope with potential changes of the socket IP during a replication. And in our case, during a restart.

(cherry picked from commit 7aa24e1c1973fb56da2d03aceaf843b71e87c8b0)