[core] fix printing cfg file context for errors after a reload_routes
Reset the file context list each time a new cfg parsing starts, to be sure we do not inherite the context from a previous parsing.
If we do not do this, when doing a reload_routes, if an parsing error occurs, we will get references and dumps (of the faulty lines) from the original, startup script, not from the reloaded one.