drouting: Allow 64-bit integers for dr_rules.sort_profile
This allows SQL views to represent the NULL value as a 64-bit integer
without getting rejected by OpenSIPS, which expected a 32-bit integer
for that column.

Admittedly, any 64-bit integer "sort profile" will now be truncated to
32 bits, but OpenSIPS had no full support for 64-bit sort profile values
in the first place, so it's better to just truncate and NOT error.