compression module: mc_compact() - Added flag `n`
Added flags param which accepts `n` for no compact headers.
When specified, opensips will only remove headers not in the whitelist
without using the compact form for From/To/CallID etc.
`mc_compact("headers|whitlelist", "n")`

The motive for this is, because some UAs do not understand the compact form
I've also encountered weird bugs using the compact form with some UAs.

This still allows for the option to whitelist headers and compact the message.